moved all input-types to input-types.ts; updated typing of AddResolvable
