+++ src/createBuild/index.ts
@@ -68,7 +68,7 @@ export const createBuild = <BuildMode = null, Context = any>(
       }),
   )
 
-  const builder: Build<BuildMode> = {
+  const builder: Build<BuildMode, Context> = {
     type: GQLBUILDER,
     buildMode,
     addQuery: createResolvable('Query'),
