--- src/input-types/create.ts
@@ -1,6 +1,0 @@
-import { createInputType } from '../utils'
-
-export const create = createInputType(
-  'create',
-  buildModeModel => buildModeModel.names.types.createType,
-)
