--- src/input-types/data.ts
@@ -1,6 +1,0 @@
-import { createInputType } from '../utils'
-
-export const data = createInputType(
-  'data',
-  buildModeModel => buildModeModel.names.types.dataType,
-)
