--- src/input-types/filter.ts
@@ -1,6 +1,0 @@
-import { createInputType } from '../utils'
-
-export const filter = createInputType(
-  'filter',
-  buildModeModel => buildModeModel.names.types.filterType,
-)
