--- src/input-types/index.ts
@@ -1,6 +1,0 @@
-export * from './create'
-export * from './data'
-export * from './filter'
-export * from './order'
-export * from './page'
-export * from './where'
