--- src/input-types/page.ts
@@ -1,6 +1,0 @@
-import { createInputType } from '../utils'
-
-export const page = createInputType(
-  'page',
-  buildModeModel => buildModeModel.names.types.pageType,
-)
