--- src/input-types/where.ts
@@ -1,6 +1,0 @@
-import { createInputType } from '../utils'
-
-export const where = createInputType(
-  'where',
-  buildModeModel => buildModeModel.names.types.whereType,
-)
